pyproject.toml
[project]
name = "calories-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.49.0",
    "streamlit>=1.51.0",
]
